Completion of dynamic forms concept
Moved to an initial form which specifies the ticket's priority and issue and
changed the rendering to render things properly. Now the user can decide
where priority shows on the client side, and the priority privacy setting is
placed in the dynamic form wizard.

The standard form is added to every ticket without option. Extra forms can
be defined and associated with help topics which can additionally be added
to tickets upon creation. This allows for standardization of the dynamic
data location for searches and filtering.

Implemented advanced search for dynamic data. Along with reinstating the
basic ticket search on keywords

Implemented ticket filtering on dynamic data for both keyword searches as
well as searches for special fields (drop-down lists, etc.)

Phone number for users is now completely optional

+++ scp/forms.php
@@ -18,20 +18,24 @@ if($_POST) {
                 $form->save();
             foreach ($form->getDynamicFields() as $field) {
                 $id = $field->get('id');
-                if ($_POST["delete-$id"] == 'on') {
+                if ($_POST["delete-$id"] == 'on' && $field->isDeletable()) {
                     $field->delete();
+                    // Don't bother updating the field
                     continue;
                 }
-                foreach (array('sort','label','type','name') as $f)
-                    if (isset($_POST["$f-$id"]))
-                        $field->set($f, $_POST["$f-$id"]);
+                if (isset($_POST["type-$id"]) && $field->isChangeable())
+                    $field->set('type', $_POST["type-$id"]);
+                if (isset($_POST["name-$id"]) && $field->isNameEditable())
+                    $field->set('name', $_POST["name-$id"]);
                 # TODO: make sure all help topics still have all required fields
-                $field->set('required', $_POST["required-$id"] == 'on' ?  1 : 0);
-                $field->set('private', $_POST["private-$id"] == 'on' ?  1 : 0);
-                # Core fields are forced required and public
-                if (in_array($field->get('name'), $required)) {
-                    $field->set('required', 1);
-                    $field->set('private', 0);
+                if (!$field->isRequirementForced())
+                    $field->set('required', $_POST["required-$id"] == 'on' ?  1 : 0);
+                if (!$field->isPrivacyForced())
+                    $field->set('private', $_POST["private-$id"] == 'on' ?  1 : 0);
+                foreach (array('sort','label') as $f) {
+                    if (isset($_POST["$f-$id"])) {
+                        $field->set($f, $_POST["$f-$id"]);
+                    }
                 }
                 if ($field->isValid())
                     $field->save();
